Use doubles instead of floats
All of the properties are doubles and we upcast when sending events
anyways, so just make everything doubles

+++ src/uzbl-core.h
@@ -132,7 +132,7 @@ typedef struct {
 
     gboolean print_version;
 
-    gfloat zoom_step;
+    gdouble zoom_step;
 
     /* command list: (key)name -> (value)Command  */
     GHashTable *commands;
